Ignore redundant/invalid alive packets.
ReacTIVision periodically sends a packet that says nothing is alive,
regardless of whether or not this is true (e.g. fingers go away then
come back again).  These packets have sequence ID -1, so are redundant
and can be ignored.  It seems like a bug to be sending invalid packets
whatsoever (some form of keepalive?) and I can't find anything about
this in the spec, but ignoring them makes the "alive noise" go away,
at least.