Fix chaotic motion on zoom and other circumstances
Subtle problem: zoom clearly made the physics go berzerk, but hand
positions and zoom scale factors have nothing to do with the map zoom
level (everything there is done in TUIO coordinates).  The reason
everything goes crazy is that zoom takes time, which means the next
iteration of the simulation will be run for a long duration, and too
long a duration is too coarse and breaks the physical model.

To fix this, always run iterations of the simulation for a fixed small
duration.  If the time between animation ticks is large, then run many
iterations; i.e. decouple the duration of a simulation tick from
real time entirely.