n2.wiki – Horizontal Rule wiki-markup fix
After testing when a horizontal rule is created with wiki-markup in
wikipedia, <hr> appear to only be generated when 4 or more dash
characters are placed independently on a new line, with no additional
characters added before or after the series of dashes.

In the following examples, only the first two examples should be
converted to <hr> elements; 

-----

----

---

--

-

     ----

a----b

Before this fix, the first 5 examples above would be converted into <hr>
elements.