Fixed but with floating label truncation with MDC Select Components
**Issue:**
When the MDC Select components have either no select menu options, or
when the options are shorter in length than the floating label's length.
The floating label is truncated, causing the label to be partially
hidden to the user. This fix, forces a min-width that will always allow
the floating label to be shown.